Removed dup_free_text. It is a duplicate of free_text. Made all other changes needed for the new component to work everywhere.
