+++ assets/src/components/v2/departures/notice_row.tsx
@@ -4,7 +4,7 @@ import FreeText from "Components/v2/free_text";
 const NoticeRow = ({ row }) => {
   return (
     <div className="departures__notice-row">
-      <FreeText elements={row.text.text} />
+      <FreeText lines={{ text: row.text.text }} />
     </div>
   );
 };
