+++ assets/src/components/v2/dup/partial_alert.tsx
@@ -1,16 +1,16 @@
 import React from "react";
 
 import { classWithModifier } from "Util/util";
-import FreeText, { FreeTextType } from "./dup_free_text";
+import FreeText, { FreeTextType } from "Components/v2/free_text";
 
 interface PartialAlertProps {
-  text: FreeTextType,
-  color: string
+  text: FreeTextType;
+  color: string;
 }
 
 const PartialAlert = (alert: PartialAlertProps) => {
-  const {text, color} = alert
-  
+  const { text, color } = alert;
+
   return (
     <div className={classWithModifier("partial-alert", color)}>
       <FreeText lines={text} />
