+++ assets/src/components/v2/free_text.tsx
@@ -3,7 +3,8 @@ import _ from "lodash";
 
 import { classWithModifier, classWithModifiers, imagePath } from "Util/util";
 
-const pillIcons = ["red", "blue", "orange", "green", "silver"];
+const textPills = ["red", "blue", "orange", "green", "silver"];
+const iconPills = ["cr", "bus"];
 
 const iconPaths: { [key: string]: string } = _.mapValues(
   {
@@ -18,16 +19,16 @@ const iconPaths: { [key: string]: string } = _.mapValues(
     green_c: "gl-c-color.svg",
     green_d: "gl-d-color.svg",
     green_e: "gl-e-color.svg",
-    clock: "clock.svg",
+    bus: "bus-black.svg",
   },
   imagePath
 );
 
-const srcForIcon = (icon) => {
+const srcForIcon = (icon: string) => {
   return iconPaths[icon];
 };
 
-const getKey = (elt) => {
+const getKey = (elt: string | FreeTextElementType) => {
   if (typeof elt === "string") {
     return elt;
   } else if (elt.format !== undefined) {
@@ -43,13 +44,15 @@ const getKey = (elt) => {
   }
 };
 
-const Icon = ({ icon }) => {
+const Icon = ({ icon }: { icon: string }) => {
   let iconElt;
 
   if (icon === null) {
     iconElt = null;
-  } else if (pillIcons.includes(icon)) {
-    iconElt = <RoutePill route={icon} />;
+  } else if (textPills.includes(icon)) {
+    iconElt = <TextRoutePill route={icon} />;
+  } else if (iconPills.includes(icon)) {
+    iconElt = <IconRoutePill route={icon} />;
   } else {
     iconElt = <img className="free-text__icon-image" src={srcForIcon(icon)} />;
   }
@@ -57,7 +60,7 @@ const Icon = ({ icon }) => {
   return <div className="free-text__icon-container">{iconElt}</div>;
 };
 
-const InlineIcon = ({ icon }) => {
+const InlineIcon = ({ icon }: { icon: string }) => {
   return (
     <span className="free-text__element free-text__inline-icon">
       <img className="free-text__inline-icon-image" src={srcForIcon(icon)} />
@@ -65,7 +68,7 @@ const InlineIcon = ({ icon }) => {
   );
 };
 
-const FormatString = ({ format, text }) => {
+const FormatString = ({ format, text }: { format: string; text: string }) => {
   const modifiers = format === null ? [] : [format];
   const className = `free-text__element ${classWithModifiers(
     "free-text__string",
@@ -75,14 +78,13 @@ const FormatString = ({ format, text }) => {
   return <span className={className}>{text}</span>;
 };
 
-const RoutePill = ({ route }) => {
+const TextRoutePill = ({ route }: { route: string }) => {
   const routeName = {
     red: "RL",
     blue: "BL",
     orange: "OL",
     green: "GL",
     silver: "SL",
-    cr: "CR",
     green_b: "GL·B",
     green_c: "GL·C",
     green_d: "GL·D",
@@ -104,7 +106,17 @@ const RoutePill = ({ route }) => {
   );
 };
 
-const TextPill = ({ color, text }) => {
+const IconRoutePill = ({ route }: { route: string }) => {
+  return (
+    <span className="free-text__element free-text__route-container">
+      <div className={classWithModifier("free-text__route-pill", route)}>
+        <img className="free-text__icon-image" src={srcForIcon(route)} />
+      </div>
+    </span>
+  );
+};
+
+const TextPill = ({ color, text }: { color: string; text: string }) => {
   return (
     <span className="free-text__element free-text__pill-container">
       <div className={classWithModifier("free-text__text-pill", color)}>
@@ -114,7 +126,7 @@ const TextPill = ({ color, text }) => {
   );
 };
 
-const Special = ({ data }) => {
+const Special = ({ data }: { data: string }) => {
   if (data === "break") {
     return <br />;
   }
@@ -122,13 +134,22 @@ const Special = ({ data }) => {
   return null;
 };
 
-const FreeTextElement = ({ elt }) => {
+interface FreeTextElementType {
+  text?: string;
+  format?: string;
+  route?: string;
+  color?: string;
+  special?: string;
+  icon?: string;
+}
+
+const FreeTextElement = ({ elt }: { elt: string | FreeTextElementType }) => {
   if (typeof elt === "string") {
     return <FormatString text={elt} format={null} />;
   } else if (elt.format !== undefined) {
     return <FormatString text={elt.text} format={elt.format} />;
   } else if (elt.route !== undefined) {
-    return <RoutePill route={elt.route} />;
+    return <TextRoutePill route={elt.route} />;
   } else if (elt.color !== undefined) {
     return <TextPill color={elt.color} text={elt.text} />;
   } else if (elt.special !== undefined) {
@@ -140,14 +161,51 @@ const FreeTextElement = ({ elt }) => {
   return null;
 };
 
-const FreeText = ({ elements }) => {
+const FreeTextLine = ({
+  icon,
+  text,
+}: {
+  icon?: string;
+  text: (string | FreeTextElementType)[];
+}) => {
   return (
-    <div className="free-text">
-      {elements.map((elt) => (
-        <FreeTextElement elt={elt} key={getKey(elt)} />
-      ))}
+    <div className="free-text__line-container">
+      {icon && <Icon icon={icon} />}
+      <div className="free-text__line">
+        {text.map((elt: string | FreeTextElementType) => (
+          <FreeTextElement elt={elt} key={getKey(elt)} />
+        ))}
+      </div>
     </div>
   );
+};
+
+export interface FreeTextType {
+  icon?: string;
+  text: FreeTextElementType[];
+}
+
+interface FreeTextProps {
+  lines: FreeTextType | FreeTextType[];
+}
+
+const FreeText = ({ lines }: FreeTextProps) => {
+  if (Array.isArray(lines)) {
+    const [{ icon: icon1, text: text1 }, { icon: icon2, text: text2 }] = lines;
+    return (
+      <div className="free-text">
+        <FreeTextLine icon={icon1} text={text1} />
+        <FreeTextLine icon={icon2} text={text2} />
+      </div>
+    );
+  } else {
+    const { icon, text } = lines;
+    return (
+      <div className="free-text">
+        <FreeTextLine icon={icon} text={text} />
+      </div>
+    );
+  }
 };
 
 export default FreeText;
